rename(UserEntity): Move Entity files to it's specific domain folder
